Adapt FDISK to the new driver structure.
- Remove all references to LUNs.
- Support up to 64 devices (with pagination).
- Devices that exist but can't be partitiones (because e.g.
  they are offline or read-only) are listed too.

Also:

- FDISK now fits in segment 5, segment 6 is completely empty.
- Removed "Test device access" (needed for reducing program size).
- Added a warning when executing _FDISK with a BASIC program in memory.
