--- source/kernel/bank5/fdisk.lnk
@@ -1,11 +1,0 @@
--mjx
--i fdisk.ihx
--b _CODE = 0x4120
--b _DATA = 0x8020
--k C:\Program Files (x86)\SDCC\bin\..\lib\z80
--l msxchar.lib
--l z80
-fdisk.rel
-fdisk_crt0.rel
-
--e
